+++ R/trajectoryRMA.R
@@ -93,7 +93,7 @@
 #' 
 #' \enc{Šidák}{Sidak}, Z. (1967) Rectangular confidence regions for the means of multivariate normal distributions. Journal of the American Statistical Association 62:648-633.
 #'
-#' @seealso \code{\link{trajectoryConvergence}}, \code{\link{trajectoryCorrespondence}}, \code{\link{trajectoryConvergencePlot}} 
+#' @seealso \code{\link{trajectoryConvergence}}, \code{\link{trajectoryCorrespondence}}, \code{\link{trajectoryConvergencePlot}}, \code{\link{trajectoryRMAPlot}} 
 #'
 #' @examples
 #' #Obtain and format some trajectories