+++ R/trajectoryComparison.R
@@ -39,6 +39,8 @@
 #'     \item{\code{TSPD}: Time-Sensitive Path Distance (experimental).}
 #'  }
 #'  
+#'  When using \code{trajectoryDistances} on trajectory cycles, then the elements to be compared are cycles (using internal states only). In this case, if TSPD is used cycle dates are used for time comparison, instead of trajectory absolute times. 
+#'  
 #'  Function \code{trajectoryConvergence} is used to study convergence/divergence between trajectories. There are three possible tests, the first two concerning pairwise comparisons between trajectories.
 #'  \enumerate{
 #'     \item{If \code{type = "pairwise.asymmetric"} then all pairwise comparisons are considered and the test is asymmetric, meaning that we test for trajectory A approaching trajectory B along time. 
@@ -271,7 +273,10 @@ trajectoryDistances<-function(x, distance.type="DSPD", symmetrization = "mean" ,
   if(inherits(x, "fd.trajectories")) {
     sites <- x$metadata$fdT
   } else if(inherits(x, "cycles")) {
-    sites <- x$metadata$cycles
+    surveys <- x$metadata$surveys[x$metadata$internal]
+    sites <- x$metadata$cycles[x$metadata$internal]
+    times <- x$metadata$dates[x$metadata$internal]
+    d <- as.dist(as.matrix(x$d)[x$metadata$internal, x$metadata$internal])
   } else if(inherits(x, "sections")) {
     sites <- x$metadata$sections
   } else {
@@ -285,8 +290,7 @@ trajectoryDistances<-function(x, distance.type="DSPD", symmetrization = "mean" ,
   for(i in 1:nsite) nsurveysite[i] = sum(sites==siteIDs[i])
   if(sum(nsurveysite==1)>0) stop("All sites need to be surveyed at least twice")
   n <- nrow(as.matrix(d))
-  nseg <- sum(nsurveysite)-nsite
-  
+
   #Init output
   dtraj <- matrix(NA, nrow=nsite, ncol = nsite)
   rownames(dtraj) <- siteIDs
@@ -446,7 +450,6 @@ trajectoryDistances<-function(x, distance.type="DSPD", symmetrization = "mean" ,
               t_high <- min(t_comp2[sel_heq])
               i_high <- which(t_comp2==t_high)
             }
-            
             if(!is.na(i_low) & !is.na(i_high)) {
               if(i_low != i_high) {
                 p <- (t_comp1[j] - t_low)/(t_high - t_low)