+++ NEWS.md
@@ -1,3 +1,6 @@
+# ecotraj 1.2.1
+* When using trajectoryDistances cycles the elements to be compared are cycles and cycle dates are used for time comparison. 
+
 # ecotraj 1.2.0
 * New functions trajectoryConvergencePlot() and cycleShiftArrows() by N. Djeghri
 * New functions trajectoryCorrespondence(), trajectoryRMA() and trajectoryRMAPlot() by N. Djeghri